Add restriction to delete admin

+++ Account Service (Java)/task/src/account/controller/AdminController.java
@@ -1,13 +1,16 @@
 package account.controller;
 
+import account.model.ROLE;
 import account.model.dto.DeleteUserResponse;
 import account.model.dto.SingUpDTO;
 import account.model.entity.AppUser;
 import account.service.UserService;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Pattern;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -32,6 +35,11 @@ public class AdminController {
     @DeleteMapping("/user/{email}")
     public ResponseEntity<DeleteUserResponse> deleteUser(@PathVariable @Email @Pattern(regexp = "^.+@acme\\.com$") String email) {
         AppUser user = userService.findUserByEmail(email);
+
+        if (user.getUserGroups().stream().anyMatch(group -> group.getName().equals(ROLE.ADMINISTRATOR))) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Can't remove ADMINISTRATOR role!");
+        }
+
         userService.deleteUser(user);
         return ResponseEntity.ok(new DeleteUserResponse(email));
     }
